requirements.txt
python-telegram-bot==21.9
openai>=1.30.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
python-dotenv>=1.0.0
